Add java buildpack for building hubspot maven artifacts, with patches to get build working
Java v2-single-module buildpack does not build due to some issues resolving test-jars. As such
we need to use the normal `v2` buildpack, but we have prereq `before` steps that need to run for every
submodule. In order to facilitate that I've created a copy_blazar_yaml.sh script which will sync the
root .blazar.yaml with all submodules.

If we ever make a change to the root .blazar.yaml, or add a new submodule, we should rerun copy_blazar_yaml.sh.
Ideally this might be made more maintainable if PaaS Build is able to tackle https://git.hubteam.com/HubSpot/PaaS-Build-Deploy/issues/1564